refactor(ssrf): single-source the redirect-hop policy in security::ssrf
Move RedirectDecision, redirect_decision, and MAX_REDIRECT_HOPS out of the
OpenAPI converter and into security::ssrf, beside validate_url_not_ssrf which
they call. Both SSRF-guarded fetch paths — openapi::convert_url and the Web-UI
spec importer fetch_spec — now consume the one shared policy instead of each
re-implementing the redirect closure, so the fail-closed behavior cannot drift
between them. fetch_spec's inline closure is replaced with the shared match;
the three redirect_decision unit tests move with the code.

No behavior change: MAX_REDIRECT_HOPS stays 5, every hop is re-validated
against the SSRF deny list, and the >=5-hop stop is unchanged.
