+++ src/capability/openapi/tests.rs
@@ -475,7 +475,7 @@ fn names(caps: &[GeneratedCapability]) -> Vec<&str> {
 
 #[test]
 fn redirect_decision_blocks_ssrf_hop() {
-    use super::convert::{RedirectDecision, redirect_decision};
+    use crate::security::ssrf::{RedirectDecision, redirect_decision};
     // A redirect hop to the cloud-metadata endpoint must be blocked, never
     // followed - closes the open-redirect SSRF vector on convert_url.
     assert!(matches!(
@@ -486,7 +486,7 @@ fn redirect_decision_blocks_ssrf_hop() {
 
 #[test]
 fn redirect_decision_follows_public_hop() {
-    use super::convert::{RedirectDecision, redirect_decision};
+    use crate::security::ssrf::{RedirectDecision, redirect_decision};
     // 1.1.1.1 is a public address literal: passes the SSRF check without DNS.
     assert_eq!(
         redirect_decision(0, "https://1.1.1.1/openapi.json"),
@@ -496,7 +496,7 @@ fn redirect_decision_follows_public_hop() {
 
 #[test]
 fn redirect_decision_stops_after_max_hops() {
-    use super::convert::{RedirectDecision, redirect_decision};
+    use crate::security::ssrf::{RedirectDecision, redirect_decision};
     // Fail-closed: once the hop budget is exhausted, stop regardless of target.
     assert_eq!(
         redirect_decision(5, "https://1.1.1.1/openapi.json"),
